functions_youtube/requirements.txt
firebase_functions~=0.5.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.18.0
requests>=2.28.0
python-dotenv>=1.0.0
google-genai>=1.0.0
youtube-transcript-api>=0.6.0